+++ src/modules/transfers/transfers.service.ts
@@ -1,15 +1,17 @@
 import { Request } from 'express';
+import { Server as IO } from 'socket.io';
 
-import SqlTransferService from '../sql/sql-transfer.service';
-import ApiTransferService from '../api/api-transfer.service';
-import ResponseHandler from '../../utils/response-handler/response-handler';
+import { SqlTransferService } from '../sql/sql-transfer.service';
+import { ApiTransferService } from '../api/api-transfer.service';
+import { ResponseHandler } from '../../utils/response-handler/response-handler';
 import { TransferStatus } from './enums/transfer-status.enum';
-import TransfersRepository from './transfers.repository';
-import ProcessesRepository from '../processes/process.repository';
+import { TransfersRepository } from './transfers.repository';
+import { ProcessesRepository } from '../processes/process.repository';
 import { Source } from '../imports/enums/source.enum';
-import EmailTransferService from '../email/email-transfer.service';
+import { EmailTransferService } from '../email/email-transfer.service';
 
-class TransfersService {
+export class TransfersService {
+  private io: IO;
   private sqlTransferService: SqlTransferService;
   private apiTransferService: ApiTransferService;
   private emailTransferService: EmailTransferService;
@@ -17,12 +19,14 @@ class TransfersService {
   private processesRepository: ProcessesRepository;
 
   constructor(
+    io: IO,
     sqlTransferService: SqlTransferService,
     apiTransferService: ApiTransferService,
     emailTransferService: EmailTransferService,
     transfersRepository: TransfersRepository,
     processesRepository: ProcessesRepository
   ) {
+    this.io = io;
     this.sqlTransferService = sqlTransferService;
     this.apiTransferService = apiTransferService;
     this.emailTransferService = emailTransferService;
@@ -49,19 +53,22 @@ class TransfersService {
   async delete(id: number) {
     const responseHandler = new ResponseHandler();
     try {
-      // const transfer = await this.transfersRepository.load(id);
-      // if (transfer === undefined) {
-      //   responseHandler.setError(404, 'Transfer not found');
-      //   return responseHandler;
-      // }
+      const transfer = await this.transfersRepository.load(id);
+      if (transfer === undefined) {
+        responseHandler.setError(404, 'Transfer not found');
+        return responseHandler;
+      }
 
-      // if (transfer.status === TransferStatus.PENDING) {
-      //   responseHandler.setError(
-      //     409,
-      //     'Pending import process cannot be deleted'
-      //   );
-      //   return responseHandler;
-      // }
+      if (
+        transfer.status === TransferStatus.PENDING ||
+        transfer.status === TransferStatus.PAUSING
+      ) {
+        responseHandler.setError(
+          409,
+          'Pending import process cannot be deleted'
+        );
+        return responseHandler;
+      }
 
       await this.transfersRepository.delete(id);
       responseHandler.setSuccess(200, true);
@@ -76,6 +83,7 @@ class TransfersService {
     const responseHandler = new ResponseHandler();
     try {
       const transfer = await this.transfersRepository.load(id);
+      const { id: unitId } = transfer.__.inUnit;
       if (transfer === undefined) {
         responseHandler.setError(404, 'Transfer not found');
         return responseHandler;
@@ -89,11 +97,13 @@ class TransfersService {
         return responseHandler;
       }
 
-      await this.transfersRepository.update({
+      const pausingTransfer = await this.transfersRepository.update({
         id,
-        status: TransferStatus.PAUSED,
+        status: TransferStatus.PAUSING,
         log: 'Transfer was pused'
       });
+
+      this.io.to(String(unitId)).emit('transfer', pausingTransfer);
       responseHandler.setSuccess(200, true);
       return responseHandler;
     } catch (error) {
@@ -122,7 +132,6 @@ class TransfersService {
         return responseHandler;
       }
 
-      // console.log('transfer.__: ', transfer.__)
       const { id: importId } = transfer.__.inImport;
       const impt = await this.processesRepository.load(importId);
       if (impt === undefined) {
@@ -130,37 +139,37 @@ class TransfersService {
         return responseHandler;
       }
 
-      // const { id: unitId } = transfer.__.inUnit;
+      const { id: unitId } = transfer.__.inUnit;
 
-      // const pendingUnitTransfer = await this.transfersRepository.query(
-      //   {
-      //     operator: 'and',
-      //     conditions: [
-      //       {
-      //         type: 'equals',
-      //         property: 'status',
-      //         value: TransferStatus.PENDING
-      //       },
-      //       {
-      //         type: 'hasEdge',
-      //         direction: 'out',
-      //         label: 'inUnit',
-      //         value: unitId
-      //       }
-      //     ]
-      //   },
-      //   {},
-      //   true
-      // );
+      const pendingUnitTransfer = await this.transfersRepository.query(
+        {
+          operator: 'and',
+          conditions: [
+            {
+              type: 'equals',
+              property: 'status',
+              value: TransferStatus.PENDING
+            },
+            {
+              type: 'hasEdge',
+              direction: 'in',
+              label: 'inUnit',
+              value: unitId
+            }
+          ]
+        },
+        {},
+        true
+      );
 
-      // console.log('pendingUnitTransfer: ', pendingUnitTransfer);
-      // if (pendingUnitTransfer) {
-      //   responseHandler.setError(
-      //     409,
-      //     'This unit is already processing another transfer.'
-      //   );
-      //   return responseHandler;
-      // }
+      console.log('pendingUnitTransfer: ', pendingUnitTransfer);
+      if (pendingUnitTransfer) {
+        responseHandler.setError(
+          409,
+          'This unit is already processing another transfer.'
+        );
+        return responseHandler;
+      }
 
       const updatedTransfer = await this.transfersRepository.update({
         id,
@@ -168,6 +177,8 @@ class TransfersService {
         retryAttempts: 0,
         log: 'Transfer was reloaded'
       });
+
+      this.io.to(String(unitId)).emit('transfer', updatedTransfer);
 
       const { source } = impt;
 
@@ -208,16 +219,16 @@ class TransfersService {
         return responseHandler;
       }
 
-      if (
-        transfer.status !== TransferStatus.PAUSED &&
-        transfer.status !== TransferStatus.FAILED
-      ) {
-        responseHandler.setError(
-          409,
-          'Only paused or failed transfer can be restarted'
-        );
-        return responseHandler;
-      }
+      // if (
+      //   transfer.status !== TransferStatus.PAUSED &&
+      //   transfer.status !== TransferStatus.FAILED
+      // ) {
+      //   responseHandler.setError(
+      //     409,
+      //     'Only paused or failed transfer can be restarted'
+      //   );
+      //   return responseHandler;
+      // }
 
       const { id: importId } = transfer.__.inImport;
       const impt = await this.processesRepository.load(importId);
@@ -228,48 +239,53 @@ class TransfersService {
 
       const { id: unitId } = transfer.__.inUnit;
 
-      // const pendingUnitTransfer = await this.transfersRepository.query(
-      //   {
-      //     operator: 'and',
-      //     conditions: [
-      //       {
-      //         type: 'equals',
-      //         property: 'status',
-      //         value: TransferStatus.PENDING
-      //       },
-      //       {
-      //         type: 'hasEdge',
-      //         direction: 'out',
-      //         label: 'inUnit',
-      //         value: unitId
-      //       }
-      //     ]
-      //   },
-      //   {},
-      //   true
-      // );
-      // if (pendingUnitTransfer) {
-      //   responseHandler.setError(
-      //     409,
-      //     'This unit is already processing another transfer.'
-      //   );
-      //   return responseHandler;
-      // }
+      const pendingUnitTransfer = await this.transfersRepository.query(
+        {
+          operator: 'and',
+          conditions: [
+            {
+              type: 'equals',
+              property: 'status',
+              value: TransferStatus.PENDING
+            },
+            {
+              type: 'hasEdge',
+              direction: 'in',
+              label: 'inUnit',
+              value: unitId
+            }
+          ]
+        },
+        {},
+        true
+      );
+      if (pendingUnitTransfer) {
+        responseHandler.setError(
+          409,
+          'This unit is already processing another transfer.'
+        );
+        return responseHandler;
+      }
 
-      const updatedTransfer = await this.transfersRepository.update({
+      let updatedTransfer = await this.transfersRepository.update({
         id,
         status: TransferStatus.PENDING,
+        references: undefined,
         offset: 0,
         transferedDatasetsCount: 0,
+        log: 'Transfer was restarted',
         retryAttempts: 0
       });
+
+      this.io.to(String(unitId)).emit('transfer', updatedTransfer);
 
       const { source } = impt;
 
       switch (source) {
         case Source.SQL: {
-          return await this.sqlTransferService.retry(impt, updatedTransfer);
+          return await this.sqlTransferService.restart(impt, updatedTransfer);
         }
+        //fix later
         case Source.API: {
           return await this.apiTransferService.retry(
             req,
@@ -278,7 +294,7 @@ class TransfersService {
           );
         }
         case Source.EMAIL: {
-          return await this.emailTransferService.retry(impt, updatedTransfer);
+          return await this.emailTransferService.restart(impt, updatedTransfer);
         }
         default: {
           responseHandler.setError(
@@ -294,5 +310,3 @@ class TransfersService {
     }
   }
 }
-
-export default TransfersService;
